lots of work on ProxyMapper.
ProxyMapper will allow flows to handle objects in a form other than the flattened out form today.

Right now the flow properties are "flattened out" for example, broadcastProvider.getTagMap() is flattened into another flowProperty "providerTagMap". This results in a problem where  "broadcastProvider" flow property should not have it's broadcastProvder.getTagMap() method called.

If the broadcastProvider.getTagMap() method is called, this is a back door to change the tagmap without going through the flow properties.

ProxyMapper will help fix this problem by wrapping the BroadcastProvider and only allowing the permitted methods to be called but not exposing the real BP object.

git-svn-id: https://svn1.hosted-projects.com/transparentpolitics/amplafi-flow/amplafi-sworddance/trunk@192 6eaf30a2-012b-4626-ba8e-69dab18af4f2